chore: separate type from index.ts
